Convert all clinical data explicitly
The converter used to interpret the absence of participants as a proxy
for converting all the clinical data implicitly. This is a problem in
case the conversion of the imaging data failed because no suitable data
were found. Instead, let's make this behavior explicit using the option
to request only the clinical data be processed.