More robust fetching of participants (#420)
* More robust fetching of participants

The previous implementation fetched all directories within the BIDS path
regardless of whether their name matched the sub-* pattern. Directories
like `conversion_info` were wrongly listed as participant IDs later on.

* Fix warning for several imports

* More robust fetching of participant folders

* Move empty dataframe guard upstream

Avoid potentiel KeyError due to missing session_id. This is more of a
stop gap fix. A proper solution would require a deeper rewrite of the
processing flow.

* Convert all clinical data explicitly

The converter used to interpret the absence of participants as a proxy
for converting all the clinical data implicitly. This is a problem in
case the conversion of the imaging data failed because no suitable data
were found. Instead, let's make this behavior explicit using the option
to request only the clinical data be processed.

* Enable clinical-data-only conversion with a custom subject list

* Only process scan files if conversion info exists

* Add missing optional value for subjects list path