Monkey-patch nipype to use Python's RFH
The concurrent RFH implementation used by Nipype internally suffers from
permission issues when running Clinica on the network. Prior to version
0.5.x, Clinica used to override the nipype loggers and (accidently?)
switched the RFH implementation to use the built-in Python one.

This commit makes this switch explicit by simply monkey-patching
Nipype's logging module to use the correct RFH for the lifetime of the
Clinica process.